+++ xarray/core/types.py
@@ -318,3 +318,4 @@ GroupIndices = tuple[GroupIndex, ...]
 Bins = Union[
     int, Sequence[int], Sequence[float], Sequence[pd.Timestamp], np.ndarray, pd.Index
 ]
+ToDictDataOptions = Literal["list", "array", True, False]